- Moved v2 DTOs into separate files in a separate folder  - fixed conflicts
