+++ src/api/tests/quote-v2.controller.spec.ts
@@ -1,17 +1,16 @@
 import { createMock } from '@golevelup/ts-jest'
-import { Test, TestingModule } from '@nestjs/testing'
-import { QuoteV2Controller } from '@/api/quote-v2.controller'
-import { QuoteService } from '@/quote/quote.service'
-import { QuoteV2TransformService } from '@/quote/services/quote-v2-transform.service'
-import { QuoteV2RequestTransformService } from '@/quote/services/quote-v2-request-transform.service'
 import { EcoAnalyticsService } from '@/analytics'
-import { QuoteIntentDataDTO } from '@/quote/dto/quote.intent.data.dto'
-import { QuoteV2RequestDTO } from '@/quote/dto/quote-v2-request.dto'
-import { QuoteDataDTO } from '@/quote/dto/quote-data.dto'
-import { QuoteV2ResponseDTO } from '@/quote/dto/quote-v2-response.dto'
-import { SolverUnsupported } from '@/quote/errors'
 import { IntentExecutionType } from '@/quote/enums/intent-execution-type.enum'
-import { zeroAddress } from 'viem'
+import { QuoteDataDTO } from '@/quote/dto/quote-data.dto'
+import { QuoteIntentDataDTO } from '@/quote/dto/quote.intent.data.dto'
+import { QuoteService } from '@/quote/quote.service'
+import { QuoteV2Controller } from '@/api/quote-v2.controller'
+import { QuoteV2RequestDTO } from '@/quote/dto/v2/quote-v2-request.dto'
+import { QuoteV2RequestTransformService } from '@/quote/services/quote-v2-request-transform.service'
+import { QuoteV2ResponseDTO } from '@/quote/dto/v2/quote-v2-response.dto'
+import { QuoteV2TransformService } from '@/quote/services/quote-v2-transform.service'
+import { SolverUnsupported } from '@/quote/errors'
+import { Test, TestingModule } from '@nestjs/testing'
 
 describe('QuoteV2Controller', () => {
   let controller: QuoteV2Controller
